refactor: move styles.jsx to components directory
 - pages/styles.jsx -> components/styles.jsx
