+++ src/pages/Landing.jsx
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types';
 import NavBar from '../components/layouts/NavBar';
 import BranchBar from '../components/layouts/BranchBar';
 import ContentBox from '../components/layouts/ContentBox';
-import { BodyWrapper, HeaderWrapper } from './styles';
+import { BodyWrapper, HeaderWrapper } from '../components/styles';
 
 import UI from '../constants/ui';
